refactor: Rename git.py because it clashes with git module
Python 2 prefixes python path with the current directory making, the
file has to be renamed to find the real git module.
